clients/cardscan-python/pyproject.toml
[tool.poetry]
name = "cardscan-client"
version = "0.3.0"
description = "Python CardScan API client"
authors = ["Team CardScan <dev@cardscan.ai>"]
license = "SEE LICENSE IN LICENSE.md"
readme = "README.md"
repository = "https://github.com/CardScan-ai/api-clients.git"
keywords = ["CardScan", "CardScan client", "CardScan API client"]

[tool.poetry.dependencies]
python = "^3.8"

urllib3 = ">= 1.25.3"
python-dateutil = ">=2.8.2"
aiohttp = ">= 3.8.4"
pydantic = "^1.10.5, <2"
aenum = ">=3.1.11"
websockets = ">=12.0"

[tool.poetry.dev-dependencies]
pytest = ">=7.2.1"
tox = ">=3.9.0"
flake8 = ">=4.0.0"
pytest-asyncio = ">=0.23.7"
python-dotenv = ">=1.0.1"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
